Restructuration of libafl dirs. Formatting + export of maps struct and functions.
