+++ include/libafl/exit.h
