+++ include/libafl/syx-snapshot/device-save.h
