+++ include/libafl/syx-snapshot/syx-cow-cache.h
