+++ libafl/syx-snapshot/device-save.c
