+++ linux-user/cpu_loop-common.h
@@ -25,7 +25,7 @@
 
 //// --- Begin LibAFL code ---
 
-#include "libafl_extras/exit.h"
+#include "libafl/exit.h"
 
 //// --- End LibAFL code ---
